+++ source/base/guid.cc
@@ -140,7 +140,8 @@ bool Guid::isNull() const
 // static
 Guid Guid::create()
 {
-    uint64_t seed = std::chrono::high_resolution_clock::now().time_since_epoch().count();
+    uint64_t seed = static_cast<uint64_t>(
+        std::chrono::high_resolution_clock::now().time_since_epoch().count());
     std::mt19937_64 engine(seed);
 
     std::uniform_int_distribution<uint64_t> uniform_distance(
